feat: refactor config
