+++ packages/config/src/chains/const/contracts.ts
@@ -1,10 +1,17 @@
-import { SupportedChainId } from '../../chains/types'
-import { mapAddressToSupportedNetworks } from '../utils/config'
+import { mapAddressToSupportedNetworks } from './utils'
+import { AdditionalTargetChainId, ChainInfo, SupportedChainId } from '../types'
+import { mainnet } from '../details/mainnet'
+import { gnosisChain } from '../details/gnosis'
+import { arbitrumOne } from '../details/arbitrum'
+import { base } from '../details/base'
+import { sepolia } from '../details/sepolia'
+import { avalanche } from '../details/avalanche'
+import { optimism } from '../details/optimism'
+import { polygon } from '../details/polygon'
 
 export const ETH_ADDRESS = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE'
 export const EXTENSIBLE_FALLBACK_HANDLER = '0x2f55e8b20D0B9FEFA187AA7d00B6Cbe563605bF5'
 export const COMPOSABLE_COW = '0xfdaFc9d1902f4e0b84f65F49f244b32b31013b74'
-
 export const COW_SHED_FACTORY = '0x00E989b87700514118Fa55326CD1cCE82faebEF6'
 export const COW_SHED_IMPLEMENTATION = '0x2CFFA8cf11B90C9F437567b86352169dF4009F73'
 
@@ -47,3 +54,38 @@ export const BARN_ETH_FLOW_ADDRESSES: Record<SupportedChainId, string> =
   mapAddressToSupportedNetworks(BARN_ETH_FLOW_ADDRESS)
 
 export const MAX_VALID_TO_EPOCH = 4294967295 // Max uint32 (Feb 07 2106 07:28:15 GMT+0100)
+
+/**
+ * Details of all supported chains.
+ */
+export const ALL_SUPPORTED_CHAINS_MAP: Record<SupportedChainId, ChainInfo> = {
+  [SupportedChainId.MAINNET]: mainnet,
+  [SupportedChainId.GNOSIS_CHAIN]: gnosisChain,
+  [SupportedChainId.ARBITRUM_ONE]: arbitrumOne,
+  [SupportedChainId.BASE]: base,
+  [SupportedChainId.AVALANCHE]: avalanche,
+  [SupportedChainId.POLYGON]: polygon,
+  [SupportedChainId.SEPOLIA]: sepolia,
+}
+
+/**
+ * All supported chains.
+ */
+export const ALL_SUPPORTED_CHAINS = Object.values(ALL_SUPPORTED_CHAINS_MAP)
+
+/**
+ * Maps a chain where you can bridge to, but not sell tokens from (not supported by CoW Protocol)
+ */
+export const ADDITIONAL_TARGET_CHAINS_MAP: Record<AdditionalTargetChainId, ChainInfo> = {
+  [AdditionalTargetChainId.OPTIMISM]: optimism,
+}
+
+/**
+ * All chains (both supported by CoW Protocol, or chains where you can bridge to)
+ */
+export const ALL_CHAINS = ALL_SUPPORTED_CHAINS.concat(Object.values(ADDITIONAL_TARGET_CHAINS_MAP))
+
+/**
+ * All chain ids (both supported by CoW Protocol, or chains where you can bridge to)
+ */
+export const ALL_CHAINS_IDS = ALL_CHAINS.map((chain) => chain.id)
