+++ packages/config/src/chains/details/polygon.ts
@@ -1,6 +1,6 @@
-import { nativeCurrencyTemplate } from '../../common/consts/tokens'
+import { nativeCurrencyTemplate } from '../../constants/tokens'
 import { ChainInfo } from '../types'
-import { RAW_CHAINS_FILES_PATH } from '../const/path'
+import { RAW_CHAINS_FILES_PATH } from '../../constants/paths'
 
 const polygonLogo = `${RAW_CHAINS_FILES_PATH}/images/polygon-logo.svg`
