+++ packages/config/src/chains/index.ts
@@ -1,3 +1,11 @@
 export * from './types'
 export * from './const'
 export * from './utils'
+export * from './details/arbitrum'
+export * from './details/avalanche'
+export * from './details/base'
+export * from './details/gnosis'
+export * from './details/mainnet'
+export * from './details/optimism'
+export * from './details/polygon'
+export * from './details/sepolia'
