+++ packages/config/src/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './ipfs'
+export * from './paths'
