--- packages/config/src/constants/path.ts
@@ -1,1 +1,0 @@
-export const RAW_FILES_PATH = 'https://raw.githubusercontent.com/cowprotocol/cow-sdk/refs/heads/main'
