+++ packages/config/src/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Configuration constants and settings for CoW Protocol SDK
+ * @module @cowprotocol/sdk-config
+ */
+export * from './chains'
+export * from './constants'
+export * from './types'
