--- packages/config/src/types/ethereum.ts
@@ -1,5 +1,0 @@
-export interface EvmCall {
-  to: string
-  data: string
-  value: bigint
-}
